norm executable/
