+++ srcs/executable/executable_3.c
@@ -6,7 +6,7 @@
 /*   By: bsadie <bsadie@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/17 13:00:58 by eghis             #+#    #+#             */
-/*   Updated: 2021/09/09 14:19:56 by bsadie           ###   ########.fr       */
+/*   Updated: 2021/09/09 17:50:19 by bsadie           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,6 @@ void	close_wait(int status, int *fd)
 
 void	unlink_close(t_all *all, t_pipe *node)
 {
-	//unlink(node->heredoc);
 	all->fd_heredoc = 0;
 	close(node->fdout);
 	close(node->fdin);
